third_party/blink/web_tests/external/wpt/trusted-types/navigate-to-javascript-url-csp-headers-expected.txt

This is a testharness.js-based test.
[FAIL] One report-only require-trusted-types-for 'script' directive: navigation continues, violation is reported.
  assert_equals: expected 1 but got 2
[FAIL] Multiple report-only require-trusted-types-for directives: one violation reported for each require-trusted-types-for 'script', invalid sink groups ignored.
  assert_equals: expected 5 but got 10
Harness: the test ran to completion.
